remove AudioData from native modules
