+++ android/src/main/java/com/videoeditorreactnative/VideoEditorModule.kt
@@ -7,6 +7,7 @@ import android.util.Log
 import com.banuba.sdk.cameraui.data.PipConfig
 import com.banuba.sdk.core.ext.isFileUrl
 import com.banuba.sdk.core.license.BanubaVideoEditor
+import com.banuba.sdk.core.data.TrackData
 import com.banuba.sdk.export.data.ExportResult
 import com.banuba.sdk.export.utils.EXTRA_EXPORTED_SUCCESS
 import com.banuba.sdk.ve.flow.VideoCreationActivity
@@ -23,6 +24,7 @@ import org.json.JSONObject
 import org.json.JSONException
 import org.json.JSONArray
 import java.io.File
+import java.util.UUID
 
 class VideoEditorModule(reactContext: ReactApplicationContext) :
     ReactContextBaseJavaModule(reactContext) {
@@ -129,7 +131,7 @@ class VideoEditorModule(reactContext: ReactApplicationContext) :
 
         val exportData = parseExportData(inputParams.getString(INPUT_PARAM_EXPORT_DATA))
 
-        val audioData = parseAudioData(inputParams.getString(INPUT_PARAM_AUDIO_DATA)) 
+        val trackData = obtainTrackData(inputParams.getString(INPUT_PARAM_AUDIO_DATA)) 
 
         initialize(licenseToken, featuresConfig, exportData) {
             val hostActivity = currentActivity
@@ -251,7 +253,7 @@ class VideoEditorModule(reactContext: ReactApplicationContext) :
                         // setup data that will be acceptable during export flow
                         additionalExportData = null,
                         // set TrackData object if you open VideoCreationActivity with preselected music track
-                        audioTrackData = audioData?.getTrackData(),
+                        audioTrackData = trackData,
                         // set Editor video configuration
                         predefinedVideos = videoSources.map { Uri.parse(it) }
                             .toTypedArray(),
@@ -291,6 +293,24 @@ class VideoEditorModule(reactContext: ReactApplicationContext) :
         }
         return jsonArray.toString().replace("\\", "")
     }
+
+    private fun obtainTrackData(audioDataJSON: String?): TrackData? = 
+        if (audioDataJSON.isNullOrEmpty()) {
+            null
+        } else {
+            try {
+                val audioDataObject = JSONObject(audioDataJSON)
+                TrackData(
+                    id = UUID.fromString(audioDataObject.optString(AUDIO_DATA_ID)),
+                    title = audioDataObject.optString(AUDIO_DATA_TITLE),
+                    subtitle = audioDataObject.optString(AUDIO_DATA_SUBTITLE),
+                    localUri = Uri.parse(audioDataObject.optString(AUDIO_DATA_LOCAL_URL))
+                )
+            } catch (e: JSONException) {
+                Log.w(TAG, "Missing Audio Data Param", e)
+                null
+            }
+        }
 
     private fun initialize(
         token: String,
