+++ ios/VideoEditorModule.swift
@@ -141,7 +141,7 @@ class VideoEditorModule: VideoEditor {
     func openVideoEditorEditor(
         fromViewController controller: UIViewController,
         videoSources: Array<URL>,
-        audioData: AudioData?,
+        mediaTrack: MediaTrack?,
         _ resolve: @escaping RCTPromiseResolveBlock,
         _ reject: @escaping RCTPromiseRejectBlock
     ) {
@@ -155,7 +155,7 @@ class VideoEditorModule: VideoEditor {
             hostController: controller,
             videoItems: videoSources,
             shouldCopyVideo: true,
-            musicTrack: audioData?.getMediaTrack(),
+            musicTrack: mediaTrack,
             animated: true
         )
